Get things working with 0.14
This now uses the sbt-riddl plugin so "compile"
or "test" will automatically translate the riddl
to hugo. Generally, all you need to do now is
"sbt test". Test cases can now be added for
validating the generated output, as needed.

You can still use newer version of riddl if you
like. The 0.14.0 plugin can work with newer
versions of riddlc. To set the minimum, specify
it in build.sbt with the riddlcMinVersion setting.